ensure instructions print via debug_show
